LayoutTests/accessibility/win/multiple-select-element-role-expected.txt

This tests the accessibility roles of multiselect elements and their option elements.

PASS accessibilityController.focusedElement.role is "AXRole: AXList"
PASS accessibilityController.focusedElement.childAtIndex(0).role is "AXRole: AXStaticText"
